requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-queryfields==1.0.0
fake-factory==9999.9.9
Faker==6.5.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
openapi-codec==1.3.2
pilkit==2.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.3
